finish changes to caching and s3 utils to get attachments

+++ openpuc_scrapers/db/s3_wrapper.py
@@ -28,6 +28,7 @@ import secrets
 
 
 import logging
+import shutil
 
 
 default_logger = logging.getLogger(__name__)
@@ -116,12 +117,14 @@ class S3FileManager:
     # S3 Stuff Below this point
 
     def download_s3_file_to_path(
-        self, file_name: str, bucket: Optional[str] = None
+        self, file_name: str, bucket: Optional[str] = None, serve_cache: bool = False
     ) -> Optional[Path]:
         file_path = self.get_local_dir_from_key(file_name)
         if bucket is None:
             bucket = self.bucket
         if file_path.is_file():
+            if serve_cache:
+                return file_path
             os.remove(file_path)
             # raise Exception("File Already Present at Path, not downloading")
         try:
@@ -134,9 +137,9 @@ class S3FileManager:
             return None
 
     def download_s3_file_to_string(
-        self, file_name: str, bucket: Optional[str] = None
+        self, file_name: str, bucket: Optional[str] = None, serve_cache: bool = False
     ) -> str:
-        path = self.download_s3_file_to_path(file_name, bucket)
+        path = self.download_s3_file_to_path(file_name, bucket, serve_cache=serve_cache)
         if path is None:
             raise ValueError("Error Encountered getting file from s3.")
         with open(path, "r", encoding="utf-8") as f:
@@ -193,8 +196,14 @@ class S3FileManager:
         return self.download_file_to_path(url, savedir)
 
     def push_file_to_s3(
-        self, filepath: Path, file_upload_name: str, bucket: Optional[str] = None
+        self, filepath: Path, file_upload_key: str, bucket: Optional[str] = None
     ) -> str:
         if bucket is None:
             bucket = self.bucket
-        return self.s3.upload_file(str(filepath), bucket, file_upload_name)
+        if self.s3_cache_directory is not None:
+            try:
+                local_cache = self.get_local_dir_from_key(file_upload_key)
+                shutil.copyfile(filepath, local_cache)
+            except Exception as e:
+                default_logger.warning(f"Encountered error copying file to cache: {e}")
+        return self.s3.upload_file(str(filepath), bucket, file_upload_key)
